Changed instructors vectors to be a private member of scheduler. Although, I think I'm going to have this set up differently in future commits

+++ Optimizer.cpp
@@ -24,7 +24,7 @@ Optimizer::~Optimizer() {
 }
 
 
-long Optimizer::getScheduleFitness(Schedule* schedule) {
+int Optimizer::getScheduleFitness(Schedule* schedule) {
 
     // Reset fitnes value
     long fitness = 0;
